refactor/file-names: naming for recoverable_apis & remove CH file
