+++ safe_core/src/lib.rs
@@ -43,7 +43,7 @@ pub use ffi::*;
 pub mod client;
 /// Config file handling.
 pub mod config_handler;
-/// Core structs and associatied functionality
+/// Core structs and associated functionality
 pub mod core_structs;
 /// Cryptographic utilities.
 pub mod crypto;
@@ -66,8 +66,7 @@ mod errors;
 mod network_event;
 
 pub use self::client::{
-    client_handler, mdata_info, recovery_wrapped_apis, test_create_balance, AuthActions, Client,
-    ClientKeys, MDataInfo,
+    mdata_info, recoverable_apis, test_create_balance, AuthActions, Client, ClientKeys, MDataInfo,
 };
 #[cfg(feature = "mock-network")]
 pub use self::client::{mock_vault_path, MockConnectionManager as ConnectionManager};
